[skip ci]
Use EntityGraphs; these let you specify that children be
loaded up front, with one SQL statement instead of the
N + 1 problem, e.g., it loads the workflow, then it fetches
each label with a separate SQL statement.

The problem is these don't work for self-joins, so checker
workflows, which are self-joins, ignore the entity graphs.

If I make SourceFiles lazy, remove the code in TooslImplCommon
that fetches SourceFiles, only fetch all published workflows, not tools,
then the number of SQL Statements goes down from 34K to 7K. That is
a huge improvement, but it's still a lot of SQL statements.

The other approach has disadvantages too, but it will be a lot faster.